+++ packages/compiler-cli/src/ngtsc/diagnostics/src/error_details_base_url.ts
@@ -11,6 +11,6 @@
  *
  * Keep the files below in full sync:
  *  - packages/compiler-cli/src/ngtsc/diagnostics/src/error_details_base_url.ts
- *  - packages/core/src/render3/error_details_base_url.ts
+ *  - packages/core/src/error_details_base_url.ts
  */
 export const ERROR_DETAILS_PAGE_BASE_URL = 'https://angular.io/errors';
