+++ packages/core/src/util/errors.ts
@@ -6,7 +6,6 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-export const ERROR_TYPE = 'ngType';
 export const ERROR_ORIGINAL_ERROR = 'ngOriginalError';
 export const ERROR_LOGGER = 'ngErrorLogger';
 
@@ -17,4 +16,16 @@ export function wrappedError(message: string, originalError: any): Error {
   const error = Error(msg);
   (error as any)[ERROR_ORIGINAL_ERROR] = originalError;
   return error;
+}
+
+export function getOriginalError(error: Error): Error {
+  return (error as any)[ERROR_ORIGINAL_ERROR];
+}
+
+export function getErrorLogger(error: unknown): (console: Console, ...values: any[]) => void {
+  return error && (error as any)[ERROR_LOGGER] || defaultErrorLogger;
+}
+
+function defaultErrorLogger(console: Console, ...values: any[]) {
+  (<any>console.error)(...values);
 }
